# openapi.yaml
openapi: 3.1.0
info:
  title: Livepraise API
  version: 1.0.0-alpha.1
  description: |
    API HTTP do servidor Livepraise (Electron + Express).

    - **Autenticação:** `Authorization: Bearer <token>` (sessão criada em `POST /api/auth/login`).
    - **Operador local:** requisições de `localhost` podem aceder a rotas `requireOperatorAccess` sem token.
    - **WebSocket ao vivo:** `ws://{host}/ws/live` — ver tag `websocket`.
  contact:
    name: Cadimos
    url: https://github.com/cadimos/livepraise
servers:
  - url: http://localhost:3000
    description: Desenvolvimento local (porta padrão)
...
paths:
...
  /api/auth/login:
    post:
      tags: [auth]
      summary: Iniciar sessão
      operationId: authLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              operator:
                value:
                  username: admin
                  password: changeme
      responses:
        '200':
          description: Sessão criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                success:
                  value:
                    status: Sucesso
                    token: 7f3c2a1b9e8d4f6a0c1b2e3d4f5a6b7
                    user:
                      id: 1
                      username: admin
                      role: operator
                      active: true
                      createdAt: '2026-01-01T00:00:00.000Z'
                      updatedAt: '2026-01-01T00:00:00.000Z'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          description: Rate limit (20 tentativas / minuto por IP)
  /api/auth/logout:
    post:
      tags: [auth]
      summary: Terminar sessão
      operationId: authLogout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Sessão revogada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOk'
              examples:
                success:
                  value:
                    status: Sucesso
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/me:
    get:
      tags: [auth]
      summary: Utilizador da sessão actual
      operationId: authMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil público
          content:
            application/json:
              schema:
                type: object
                required: [status, user]
                properties:
                  status:
                    type: string
                    example: Sucesso
                  user:
                    $ref: '#/components/schemas/PublicUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/session:
    get:
      tags: [auth]
      summary: Validar token Bearer
      operationId: authSession
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token válido
          content:
            application/json:
              schema:
                type: object
                required: [status, user]
                properties:
                  status:
                    type: string
                  user:
                    $ref: '#/components/schemas/PublicUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /themes/{themeId}/theme.json:
    get:
      tags: [themes]
      summary: Metadados do tema
      operationId: getThemeJson
      parameters:
        - $ref: '#/components/parameters/themeId'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          description: Tema não encontrado
  /themes/{themeId}/variables.css:
    get:
      tags: [themes]
      summary: Variáveis CSS do tema
      operationId: getThemeCss
      parameters:
        - $ref: '#/components/parameters/themeId'
      responses:
        '200':
          content:
            text/css:
              schema:
                type: string
  /themes/{themeId}/assets/{filename}:
    get:
      tags: [themes]
      summary: Asset estático do tema
      operationId: getThemeAsset
      parameters:
        - $ref: '#/components/parameters/themeId'
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ficheiro binário
        '403':
          description: Path traversal bloqueado
        '404':
          description: Asset não encontrado
...
  /api/system/error-log:
    get:
      tags: [system]
      summary: Listar entradas recentes do log de erros
      operationId: listErrorLog
      security:
        - bearerAuth: []
      description: |
        Requer acesso de operador (`requireOperatorAccess`).
        Em localhost o token é opcional; remoto exige sessão operator.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 200
      responses:
        '200':
          description: Entradas mais recentes primeiro
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorLogListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [system]
      summary: Registar erro (cliente ou smoke)
      operationId: appendErrorLog
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ErrorLogAppendRequest'
            examples:
              client:
                value:
                  level: error
                  source: client
                  message: Falha ao carregar playlist
                  detail: GET /playlist → HTTP 500
      responses:
        '201':
          description: Entrada criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorLogAppendResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [system]
      summary: Limpar log de erros
      operationId: clearErrorLog
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Arquivo zerado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOk'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: session-token
      description: Token devolvido por POST /api/auth/login
  parameters:
...
    themeId:
      name: themeId
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Pedido inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Unauthorized:
      description: Autenticação obrigatória ou sessão inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: string
      example:
        error: Credenciais inválidas
    StatusOk:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: Sucesso
...
    ErrorLogEntry:
      type: object
      required: [id, ts, level, source, message]
      properties:
        id:
          type: string
          format: uuid
        ts:
          type: string
          format: date-time
        level:
          type: string
          enum: [error, warn]
        source:
          type: string
          example: api
        message:
          type: string
        detail:
          type: string
    ErrorLogListResponse:
      type: object
      required: [status, items]
      properties:
        status:
          type: string
          example: Sucesso
        items:
          type: array
          items:
            $ref: '#/components/schemas/ErrorLogEntry'
    ErrorLogAppendRequest:
      type: object
      required: [message]
      properties:
        level:
          type: string
          enum: [error, warn]
          default: error
        source:
          type: string
          default: client
        message:
          type: string
        detail:
          type: string
    ErrorLogAppendResponse:
      type: object
      required: [status, entry]
      properties:
        status:
          type: string
          example: Sucesso
        entry:
          $ref: '#/components/schemas/ErrorLogEntry'
    PublicUser:
      type: object
      required: [id, username, role, active, createdAt, updatedAt]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [operator, remote, admin]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      required: [status, token, user]
      properties:
        status:
          type: string
        token:
          type: string
        user:
          $ref: '#/components/schemas/PublicUser'
...